feat: Add top tracks section to index.html
This commit adds a new section to the index.html file, displaying the top tracks for Mexico. The section is added within a container div with the id "tops" and includes a heading with the text "Top Tracks Mexico". The track list is dynamically populated using JavaScript and will be updated based on the user's preferences.