refactor: Update user image size and alignment
This commit updates the user image size and alignment in the toptracks.js file and styleTracks.css file. The user image is now displayed with a width of 80px and height of 80px, and it is aligned properly within the user-info container. This change improves the visual appearance of the user profile section.